pyproject.toml
[project]
name = "chess-llm"
version = "0.1.0"
description = "Finetune LLMs for chess next-move prediction"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "datasets>=4.6.0",
    "huggingface-hub>=0.28.0",
    "unsloth>=2025.11.1",
    "python-chess>=1.999",
    "trl>=0.17.0",
    "wandb>=0.19.0",
    "pyyaml>=6.0",
    "tqdm>=4.67.0",
    "runpod>=1.8.1",
    "python-dotenv>=1.2.1",
    "runpod-flash>=1.5.0",
]
